rose-file-readers: Change ZmsFile::bone_indices to be vec of u32.
To be compatible with new bevy skinning implementation.

+++ rose-file-readers/src/zms.rs
@@ -24,7 +24,7 @@ pub struct ZmsFile {
     pub normal: Vec<[f32; 3]>,
     pub color: Vec<[f32; 4]>,
     pub bone_weights: Vec<[f32; 4]>,
-    pub bone_indices: Vec<[u8; 4]>,
+    pub bone_indices: Vec<[u32; 4]>,
     pub tangent: Vec<[f32; 3]>,
     pub uv1: Vec<[f32; 2]>,
     pub uv2: Vec<[f32; 2]>,
@@ -148,7 +148,7 @@ impl ZmsFile {
                     .get(index.x as usize)
                     .cloned()
                     .ok_or(ZmsReadError::InvalidBoneIndex)?;
-                bone_indices.push([bone_x as u8, bone_y as u8, bone_z as u8, bone_w as u8]);
+                bone_indices.push([bone_x as u32, bone_y as u32, bone_z as u32, bone_w as u32]);
             }
             (bone_weights, bone_indices)
         } else {
@@ -269,7 +269,7 @@ impl ZmsFile {
                     .get(index.x as usize)
                     .cloned()
                     .ok_or(ZmsReadError::InvalidBoneIndex)?;
-                bone_indices.push([bone_x as u8, bone_y as u8, bone_z as u8, bone_w as u8]);
+                bone_indices.push([bone_x as u32, bone_y as u32, bone_z as u32, bone_w as u32]);
             }
             (bone_weights, bone_indices)
         } else {
